requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bootstrap4==4.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1